rename ProductSection components according to Page.vue: CoffeeSection
