+++ presenter/src/components/sections/CoffeeSection.stories.ts
@@ -1,18 +1,18 @@
 import { SBComp } from '#types/SBComp'
 
-import FifthSection from './FifthSection.vue'
+import CoffeeSection from './CoffeeSection.vue'
 
 import type { Meta, StoryObj } from '@storybook/vue3'
 
 // More on how to set up stories at: https://storybook.js.org/docs/vue/writing-stories/introduction
 const meta = {
   title: 'Sections/Section5',
-  component: FifthSection as SBComp,
+  component: CoffeeSection as SBComp,
   // This component will have an automatically generated docsPage entry: https://storybook.js.org/docs/vue/writing-docs/autodocs
   tags: ['autodocs'],
   argTypes: {},
   args: {}, // default value
-} satisfies Meta<typeof FifthSection>
+} satisfies Meta<typeof CoffeeSection>
 
 export default meta
 type Story = StoryObj<typeof meta>
