+++ presenter/src/components/sections/CoffeeSection.vue
